+++ core/src/com/unciv/GUI.kt
@@ -77,7 +77,7 @@ object GUI {
     }
 
     private var keyboardAvailableCache: Boolean? = null
-    /** Tests availability of a physical keyboard */
+    /** Tests availability of a physical keyboard - cached (connecting a keyboard while the game is running won't be recognized until relaunch) */
     val keyboardAvailable: Boolean
         get() {
             // defer decision if Gdx.input not yet initialized