+++ core/src/com/unciv/logic/map/tile/Tile.kt
@@ -3,7 +3,6 @@
 import com.badlogic.gdx.math.Vector2
 import com.badlogic.gdx.utils.Json
 import com.badlogic.gdx.utils.JsonValue
-import com.badlogic.gdx.utils.SerializationException
 import com.unciv.Constants
 import com.unciv.GUI
 import com.unciv.UncivGame
@@ -14,7 +13,7 @@ import com.unciv.logic.civilization.Civilization
 import com.unciv.logic.civilization.PlayerType
 import com.unciv.logic.map.HexMath
 import com.unciv.logic.map.MapParameters
-import com.unciv.logic.map.MapResources
+import com.unciv.logic.map.mapgenerator.MapResourceSetting
 import com.unciv.logic.map.TileMap
 import com.unciv.logic.map.mapgenerator.MapGenerator
 import com.unciv.logic.map.mapunit.MapUnit
@@ -32,7 +31,6 @@ import com.unciv.models.ruleset.unique.UniqueType
 import com.unciv.ui.components.extensions.withItem
 import com.unciv.ui.components.extensions.withoutItem
 import com.unciv.ui.components.fonts.Fonts
-import com.unciv.ui.screens.mapeditorscreen.TileInfoNormalizer
 import com.unciv.utils.DebugUtils
 import com.unciv.utils.Log
 import kotlin.math.abs
@@ -321,12 +319,7 @@ class Tile : IsPartOfGameInfoSerialization, Json.Serializable {
         else ruleset.tileImprovements[getUnpillagedRoad().name]
     }
 
-    fun getShownImprovement(viewingCiv: Civilization?): String? {
-        return if (viewingCiv == null || viewingCiv.playerType == PlayerType.AI || viewingCiv.isSpectator())
-            improvement
-        else
-            viewingCiv.lastSeenImprovement[position]
-    }
+    fun getShownImprovement(viewingCiv: Civilization?): String? = viewingCiv?.getLastSeenImprovement(position) ?: improvement
 
     /** Returns true if this tile has fallout or an equivalent terrain feature */
     fun hasFalloutEquivalent(): Boolean = terrainFeatures.any { ruleset.terrains[it]!!.hasUnique(UniqueType.NullifyYields)}
@@ -769,8 +762,8 @@ class Tile : IsPartOfGameInfoSerialization, Json.Serializable {
         val majorDepositFinal = majorDeposit ?: (rng.nextDouble() < approximateMajorDepositDistribution())
         val depositAmounts = if (majorDepositFinal) newResource.majorDepositAmount else newResource.minorDepositAmount
         resourceAmount = when (tileMap.mapParameters.mapResources) {
-            MapResources.sparse -> depositAmounts.sparse
-            MapResources.abundant -> depositAmounts.abundant
+            MapResourceSetting.sparse.label -> depositAmounts.sparse
+            MapResourceSetting.abundant.label -> depositAmounts.abundant
             else -> depositAmounts.default
         }
     }
@@ -802,7 +795,7 @@ class Tile : IsPartOfGameInfoSerialization, Json.Serializable {
     fun setBaseTerrain(baseTerrainObject: Terrain){
         baseTerrain = baseTerrainObject.name
         this.baseTerrainObject = baseTerrainObject
-        TileInfoNormalizer.normalizeToRuleset(this, ruleset)
+        TileNormalizer.normalizeToRuleset(this, ruleset)
         setTerrainFeatures(terrainFeatures)
         setTerrainTransients()
     }