+++ core/src/com/unciv/ui/screens/devconsole/DevConsolePopup.kt
@@ -2,15 +2,26 @@ package com.unciv.ui.screens.devconsole
 
 import com.badlogic.gdx.Input
 import com.badlogic.gdx.graphics.Color
+import com.badlogic.gdx.scenes.scene2d.Actor
 import com.badlogic.gdx.scenes.scene2d.actions.Actions
+import com.badlogic.gdx.scenes.scene2d.ui.Label
+import com.badlogic.gdx.scenes.scene2d.ui.Table
+import com.badlogic.gdx.utils.Scaling
 import com.unciv.Constants
+import com.unciv.GUI
+import com.unciv.logic.civilization.Civilization
 import com.unciv.logic.map.mapunit.MapUnit
 import com.unciv.ui.components.UncivTextField
+import com.unciv.ui.components.extensions.surroundWithCircle
 import com.unciv.ui.components.extensions.toCheckBox
 import com.unciv.ui.components.extensions.toLabel
 import com.unciv.ui.components.input.KeyCharAndCode
 import com.unciv.ui.components.input.keyShortcuts
+import com.unciv.ui.components.input.onClick
+import com.unciv.ui.components.input.onRightClick
+import com.unciv.ui.images.ImageGetter
 import com.unciv.ui.popups.Popup
+import com.unciv.ui.screens.devconsole.CliInput.Companion.splitToCliInput
 import com.unciv.ui.screens.worldscreen.WorldScreen
 
 
@@ -25,6 +36,7 @@ class DevConsolePopup(val screen: WorldScreen) : Popup(screen) {
 
     private val textField = UncivTextField.create("", "") // always has focus, so a hint won't show
     private val responseLabel = "".toLabel(Color.RED).apply { wrap = true }
+    private val inputWrapper = Table()
 
     private val commandRoot = ConsoleCommandRoot()
     internal val gameInfo = screen.gameInfo
@@ -32,12 +44,16 @@ class DevConsolePopup(val screen: WorldScreen) : Popup(screen) {
     init {
         // Use untranslated text here! The entire console, including messages, should stay English.
         // But "Developer Console" *has* a translation from KeyboardBinding.DeveloperConsole.
-        // The extensions still help, even with a "don't translate" kludge ("Keep open" has no template but might in the future).
-        add("Developer Console {}".toLabel(fontSize = Constants.headingFontSize)).growX()  // translation template is automatic via the keybinding
+        // The extensions still help, even with a "don't translate" kludge
+        add("Developer Console {}".toLabel(fontSize = Constants.headingFontSize)).growX()
         add("Keep open {}".toCheckBox(keepOpen) { keepOpen = it }).right().row()
 
-        add(textField).width(stageToShowOn.width / 2).colspan(2).row()
-        textField.keyShortcuts.add(Input.Keys.ENTER, ::onEnter)
+        inputWrapper.defaults().space(5f)
+        if (!GUI.keyboardAvailable) inputWrapper.add(getAutocompleteButton())
+        inputWrapper.add(textField).growX()
+        if (!GUI.keyboardAvailable) inputWrapper.add(getHistoryButtons())
+
+        add(inputWrapper).minWidth(stageToShowOn.width / 2).growX().colspan(2).row()
 
         // Without this, console popup will always contain the key used to open it - won't work perfectly if it's configured to a "dead key"
         textField.addAction(Actions.delay(0.05f, Actions.run { textField.text = "" }))
@@ -47,10 +63,8 @@ class DevConsolePopup(val screen: WorldScreen) : Popup(screen) {
         keyShortcuts.add(KeyCharAndCode.BACK) { close() }
         clickBehindToClose = true
 
-        textField.keyShortcuts.add(KeyCharAndCode.TAB) {
-            val textToAdd = getAutocomplete()
-            textField.appendText(textToAdd)
-        }
+        textField.keyShortcuts.add(Input.Keys.ENTER, ::onEnter)
+        textField.keyShortcuts.add(KeyCharAndCode.TAB, ::onAutocomplete)
 
         keyShortcuts.add(Input.Keys.UP) { navigateHistory(-1) }
         keyShortcuts.add(Input.Keys.DOWN) { navigateHistory(1) }
@@ -60,11 +74,63 @@ class DevConsolePopup(val screen: WorldScreen) : Popup(screen) {
         screen.stage.keyboardFocus = textField
     }
 
+    private fun getAutocompleteButton() = ImageGetter.getArrowImage()
+        .surroundWithCircle(50f, color = Color.DARK_GRAY).onClick(::onAutocomplete)
+
+    private fun getHistoryButtons(): Actor {
+        val group = Table()
+        fun getArrow(rotation: Float, delta: Int) = ImageGetter.getImage("OtherIcons/ForwardArrow").apply {
+            name = if (delta > 0) "down" else "up"
+            setScaling(Scaling.fillX)
+            setSize(36f, 16f)
+            scaleX = 0.75f  // no idea why this works
+            setOrigin(18f, 8f)
+            this.rotation = rotation
+            onClick {
+                navigateHistory(delta)
+            }
+        }
+        group.add(getArrow(90f, -1)).size(36f, 16f).padBottom(4f).row()
+        group.add(getArrow(-90f, 1)).size(36f, 16f)
+        group.setSize(40f, 40f)
+        return group.surroundWithCircle(50f, false, Color.DARK_GRAY).onRightClick(action = ::showHistory)
+    }
+
+    private fun onAutocomplete() {
+        val (toRemove, toAdd) = getAutocomplete() ?: return
+        fun String.removeFromEnd(n: Int) = substring(0, (length - n).coerceAtLeast(0))
+        textField.text = textField.text.removeFromEnd(toRemove) + toAdd
+        textField.cursorPosition = Int.MAX_VALUE // because the setText implementation actively resets it after the paste it uses (auto capped at length)
+    }
+
     private fun navigateHistory(delta: Int) {
         if (history.isEmpty()) return
         currentHistoryEntry = (currentHistoryEntry + delta).coerceIn(history.indices)
         textField.text = history[currentHistoryEntry]
         textField.cursorPosition = textField.text.length
+    }
+
+    internal fun showHistory() {
+        if (history.isEmpty()) return
+        val popup = object : Popup(stageToShowOn) {
+            init {
+                for ((index, entry) in history.withIndex()) {
+                    val label = Label(entry, skin)
+                    label.onClick {
+                        currentHistoryEntry = index
+                        navigateHistory(0)
+                        close()
+                    }
+                    add(label).row()
+                }
+                clickBehindToClose = true
+                if (screen.game.settings.forbidPopupClickBehindToClose) addCloseButton()
+                showListeners.add {
+                    getScrollPane()?.run { scrollY = maxY }
+                }
+            }
+        }
+        popup.open(true)
     }
 
     private fun onEnter() {
@@ -96,30 +162,36 @@ class DevConsolePopup(val screen: WorldScreen) : Popup(screen) {
         responseLabel.style.fontColor = color
     }
 
-    private val splitStringRegex = Regex("\"([^\"]+)\"|\\S+") // Read: "(phrase)" OR non-whitespace
-    private fun getParams(text: String): List<String> {
-        return splitStringRegex.findAll(text).map { it.value.removeSurrounding("\"") }.filter { it.isNotEmpty() }.toList()
-    }
-
     private fun handleCommand(): DevConsoleResponse {
-        val params = getParams(textField.text)
+        val params = textField.text.splitToCliInput()
         return commandRoot.handle(this, params)
     }
 
-    private fun getAutocomplete(): String {
-        val params = getParams(textField.text)
-        return commandRoot.autocomplete(this, params)
+    private fun getAutocomplete(): Pair<Int,String>? {
+        val params = textField.text.splitToCliInput()
+        val autoCompleteString = commandRoot.autocomplete(this, params)
+            ?: return null
+        val replaceLength = params.lastOrNull()?.originalLength() ?: 0
+        return replaceLength to autoCompleteString
     }
 
-    internal fun getCivByName(name: String) = gameInfo.civilizations.firstOrNull { it.civName.toCliInput() == name.toCliInput() }
+    internal fun getCivByName(name: CliInput) =
+        getCivByNameOrNull(name)
         ?: throw ConsoleErrorException("Unknown civ: $name")
+    internal fun getCivByNameOrSelected(name: CliInput?) =
+        name?.let { getCivByName(it) }
+        ?: screen.selectedCiv
+    internal fun getCivByNameOrNull(name: CliInput): Civilization? =
+        gameInfo.civilizations.firstOrNull { name.equals(it.civName) }
 
-    internal fun getSelectedTile() = screen.mapHolder.selectedTile ?: throw ConsoleErrorException("Select tile")
+    internal fun getSelectedTile() = screen.mapHolder.selectedTile
+        ?: throw ConsoleErrorException("Select tile")
 
     /** Gets city by selected tile */
-    internal fun getSelectedCity() = getSelectedTile().getCity() ?: throw ConsoleErrorException("Select tile belonging to city")
+    internal fun getSelectedCity() = getSelectedTile().getCity()
+        ?: throw ConsoleErrorException("Select tile belonging to city")
 
-    internal fun getCity(cityName: String) = gameInfo.getCities().firstOrNull { it.name.toCliInput() == cityName.toCliInput() }
+    internal fun getCity(cityName: CliInput) = gameInfo.getCities().firstOrNull { cityName.equals(it.name) }
         ?: throw ConsoleErrorException("Unknown city: $cityName")
 
     internal fun getSelectedUnit(): MapUnit {
@@ -130,7 +202,4 @@ class DevConsolePopup(val screen: WorldScreen) : Popup(screen) {
         return if (selectedUnit != null && selectedUnit.getTile() == selectedTile) selectedUnit
         else units.first()
     }
-
-    internal fun getInt(param: String) = param.toIntOrNull() ?: throw ConsoleErrorException("$param is not a valid number")
-
 }